refactor: clean App and rewrite Main
