+++ src/App.css
@@ -107,8 +107,10 @@
 }
 
 .card-list-container {
+  background-color: hsla(0, 0%, 0%, 0.77);
+  padding: 20px 30px;
+  border-radius: 8px;
   width: 100%;
-  padding: 10px;
 }
 
 .card-table {
@@ -218,6 +220,7 @@
   justify-content: center;
   align-items: center;
   height: 100vh;
+  width: 100vw;
   background-color: #ffffff;
   color: rgb(251, 3, 3);
   text-align: center;
